Backport applypolicies change
