+++ tests/policy/shim.go
@@ -4,6 +4,10 @@ import "github.com/TykTechnologies/tyk/gateway"
 
 const DefaultOrg = "default-org-id"
 
-type APISpec = gateway.APISpec
+type (
+	Gateway        = gateway.Gateway
+	APISpec        = gateway.APISpec
+	BaseMiddleware = gateway.BaseMiddleware
+)
 
 var StartTest = gateway.StartTest
